adequando implementação ao modelo descritivo de desenvolvimento
