+++ listelab-contrato/Controllers/QuestaoDiscursivaController.cs
@@ -1,6 +1,7 @@
 ﻿using listelab_dominio.Conceitos;
 using listelab_dominio.Conceitos.Filtro;
-using listelab_dominio.Conceitos.Resposta;
+using listelab_dominio.Conceitos.QuestaoObj;
+using listelab_dominio.Conceitos.RespostaObj;
 using listelab_dominio.InterfaceDeServico;
 
 namespace listelab_contrato.Controllers
