--- listelab-dominio/Conceitos/Filtro/Filtro.cs
@@ -1,10 +1,0 @@
-﻿
-namespace listelab_dominio.Conceitos.Filtro
-{
-    /// <summary>
-    /// Classe abstrata que representa filtro.
-    /// </summary>
-   public abstract class Filtro
-    {
-    }
-}
