+++ listelab-dominio/Conceitos/FiltroObj/Filtro.cs
@@ -1,0 +1,10 @@
+﻿
+namespace listelab_dominio.Conceitos.Filtro
+{
+    /// <summary>
+    /// Classe abstrata que representa filtro.
+    /// </summary>
+   public abstract class Filtro
+    {
+    }
+}
