+++ listelab-dominio/Conceitos/FiltroObj/FiltroLista.cs
