+++ listelab-dominio/Conceitos/RespostaObj/Discursiva.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace listelab_dominio.Conceitos.Resposta
+namespace listelab_dominio.Conceitos.RespostaObj
 {
     public class Discursiva
     {
