+++ listelab-dominio/InterfaceDeServico/IServicoListaQuestoes.cs
@@ -1,4 +1,5 @@
 ﻿using listelab_dominio.Conceitos;
+using listelab_dominio.Conceitos.ListaObj;
 using System;
 using System.Collections.Generic;
 using System.Text;
