+++ listelab-servico/Servico/ServicoQuestaoDiscursiva.cs
@@ -2,8 +2,8 @@
 using System.Linq.Expressions;
 using listelab_data.Repositorios;
 using listelab_dominio.Conceitos.Filtro;
-using listelab_dominio.Conceitos;
-using listelab_dominio.Conceitos.Resposta;
+using listelab_dominio.Conceitos.QuestaoObj;
+using listelab_dominio.Conceitos.RespostaObj;
 using listelab_dominio.InterfaceDeServico;
 using listelab_servico.Validacoes;
