+++ listelab-servico/Servico/ServicoQuestaoObjetiva.cs
@@ -1,7 +1,7 @@
 ﻿using listelab_data.Repositorios;
-using listelab_dominio.Conceitos;
 using listelab_dominio.Conceitos.Filtro;
-using listelab_dominio.Conceitos.Resposta;
+using listelab_dominio.Conceitos.QuestaoObj;
+using listelab_dominio.Conceitos.RespostaObj;
 using listelab_dominio.InterfaceDeServico;
 using listelab_servico.Validacoes;
 using System;
