+++ listelab-servico/Validacoes/ValidacoesQuestaoObjetiva.cs
@@ -1,8 +1,6 @@
-﻿using listelab_dominio.Conceitos;
-using listelab_dominio.Conceitos.Resposta;
+﻿using listelab_dominio.Conceitos.QuestaoObj;
+using listelab_dominio.Conceitos.RespostaObj;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace listelab_servico.Validacoes
 {
